Use StaplerResponseWrapper from latest Stapler core
