+++ core/src/main/java/hudson/security/csrf/CrumbFilter.java
@@ -12,6 +12,7 @@ import org.acegisecurity.providers.anonymous.AnonymousAuthenticationToken;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.util.Collection;
 import java.util.Enumeration;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -29,6 +30,7 @@ import org.kohsuke.stapler.HttpResponseRenderer;
 import org.kohsuke.stapler.HttpResponses.HttpResponseException;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
+import org.kohsuke.stapler.StaplerResponseWrapper;
 import org.kohsuke.stapler.WebApp;
 import org.kohsuke.stapler.interceptor.RequirePOST;
 
@@ -39,6 +41,8 @@ import org.kohsuke.stapler.interceptor.RequirePOST;
  * @author dty
  */
 public class CrumbFilter implements Filter {
+    private static final String POST_PROCESSOR_EXCEPTION_CLASS_NAME = RequirePOST.Processor.class.getName() + "$1";
+
     /**
      * Because servlet containers generally don't specify the ordering of the initialization
      * (and different implementations indeed do this differently --- See HUDSON-3878),
@@ -61,9 +65,9 @@ public class CrumbFilter implements Filter {
             public boolean generateResponse(StaplerRequest req, StaplerResponse rsp, Object node, Object response) throws IOException, ServletException {
                 if (response instanceof HttpResponseException) {
                     Class<?> c = response.getClass();
-                    // The only response with enclosing class of RequirePOST.Processor is the 'try post'
-                    // response, which we need ot modify to be crumb-aware
-                    if (RequirePOST.Processor.class.equals(c.getEnclosingClass())) {
+                    // the post processor exception is the 'try post'
+                    // response, which needs to be modified to be crumb-aware
+                    if (POST_PROCESSOR_EXCEPTION_CLASS_NAME.equals(c.getName())) {
                         CrumbIssuer crumbIssuer = getCrumbIssuer();
                         if (crumbIssuer != null) {
                             String crumbFieldName = crumbIssuer.getDescriptor().getCrumbRequestField();
@@ -72,6 +76,35 @@ public class CrumbFilter implements Filter {
                             StaplerResponse res = new StaplerResponseWrapper(rsp) {
                                 public PrintWriter getWriter() throws IOException {
                                     return resWriter;
+                                }
+                                // Need extra methods since Jenkins upgraded to Servlet API 3
+                                @Override
+                                public void setContentLengthLong(long len) {
+                                    getWrapped().setContentLengthLong(len);
+                                }
+                                @Override
+                                public void setCharacterEncoding(String charset) {
+                                    getWrapped().setCharacterEncoding(charset);
+                                }
+                                @Override
+                                public String getContentType() {
+                                    return getWrapped().getContentType();
+                                }
+                                @Override
+                                public int getStatus() {
+                                    return getWrapped().getStatus();
+                                }
+                                @Override
+                                public Collection<String> getHeaders(String name) {
+                                    return getWrapped().getHeaders(name);
+                                }
+                                @Override
+                                public Collection<String> getHeaderNames() {
+                                    return getWrapped().getHeaderNames();
+                                }
+                                @Override
+                                public String getHeader(String name) {
+                                    return getWrapped().getHeader(name);
                                 }
                             };
                             ((HttpResponseException)response).generateResponse(req, res, node);
